Add Python 101 and Data Analysis 101 notebooks
These were split from the Python introduction notebook by Rick Muller.
Also included is Python 102, which is the rest of Rick Muller's sections.
This one is still Python 2 based, and has some stuff (like use of nwchem)
which we can surely replace with something more relevant.

An uncurated list of references is also added to the readme.